Create `UIViewExtensions.macios.cs` & `PopupVisualOptions.macios.cs`
