+++ src/CommunityToolkit.Maui.Core/Views/Popup/PopupView.macios.cs
@@ -2,27 +2,13 @@
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using CommunityToolkit.Maui.Core.Extensions;
+using CommunityToolkit.Maui.Core.Primitives;
 using CoreAnimation;
 using CoreGraphics;
 using UIKit;
 
 namespace CommunityToolkit.Maui.Core.Views;
-
-/// <summary>
-/// Visual Options for <see cref="PopupView"/>
-/// </summary>
-public class PopupViewVisualOptions
-{
-	/// <summary>
-	/// <see cref="PopupView"/> Border Corner Radius
-	/// </summary>
-	public CGRect CornerRadius { get; set; }
-
-	/// <summary>
-	/// <see cref="PopupView"/> Background Color
-	/// </summary>
-	public UIColor BackgroundColor { get; set; } = UIColor.Gray;
-}
 
 /// <summary>
 /// <see cref="UIView"/> for <see cref="Popup"/>
@@ -173,87 +159,4 @@ public class PopupView : UIView
 			return path.CGPath;
 		}
 	}
-}
-
-static class UIViewExtensions
-{
-	/// <summary>
-	/// Safe bottom edge of the guide
-	/// </summary>
-	public static NSLayoutYAxisAnchor SafeBottomAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.BottomAnchor
-			: view.BottomAnchor;
-
-	/// <summary>
-	/// Safe horizontal center of the guide
-	/// </summary>
-	public static NSLayoutXAxisAnchor SafeCenterXAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.CenterXAnchor
-			: view.CenterXAnchor;
-
-	/// <summary>
-	/// Safe vertical center of the guide
-	/// </summary>
-	public static NSLayoutYAxisAnchor SafeCenterYAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.CenterYAnchor
-			: view.CenterYAnchor;
-
-	/// <summary>
-	/// Safe vertical extent of the guide
-	/// </summary>
-	public static NSLayoutDimension SafeHeightAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.HeightAnchor
-			: view.HeightAnchor;
-
-	/// <summary>
-	/// Safe leading edge of the guide
-	/// </summary>
-	public static NSLayoutXAxisAnchor SafeLeadingAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.LeadingAnchor
-			: view.LeadingAnchor;
-
-	/// <summary>
-	/// Safe left edge of the guide
-	/// </summary>
-	public static NSLayoutXAxisAnchor SafeLeftAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.LeftAnchor
-			: view.LeftAnchor;
-
-	/// <summary>
-	/// Safe right edge of the guide
-	/// </summary>
-	public static NSLayoutXAxisAnchor SafeRightAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.RightAnchor
-			: view.RightAnchor;
-
-	/// <summary>
-	/// Safe top edge of the guide
-	/// </summary>
-	public static NSLayoutYAxisAnchor SafeTopAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.TopAnchor
-			: view.TopAnchor;
-
-	/// <summary>
-	/// Safe trailing edge of the guide
-	/// </summary>
-	public static NSLayoutXAxisAnchor SafeTrailingAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.TrailingAnchor
-			: view.TrailingAnchor;
-
-	/// <summary>
-	/// Safe width edge of the guide
-	/// </summary>
-	public static NSLayoutDimension SafeWidthAnchor(this UIView view) =>
-		UIDevice.CurrentDevice.CheckSystemVersion(11, 0)
-			? view.SafeAreaLayoutGuide.WidthAnchor
-			: view.WidthAnchor;
 }
